GaussFiltr v 3.0

+++ tasks/mpi/Sadikov_I_Gauss_Linear_Filtration/include/ops_mpi.h
@@ -31,7 +31,7 @@ class LinearFiltrationSeq : public ppc::core::Task {
   bool post_processing() override;
   void CalculateGaussMatrix();
   void CalculateNewPixelValue(int iIndex, int jIndex);
-  bool CheckIndex(int index) const noexcept { return index >= 0 and index < m_pixelsMatrix.size(); }
+  bool CheckIndex(int index) const noexcept { return index >= 0 and index < static_cast<int>(m_pixelsMatrix.size()); }
   bool CheckRowIndex(int index) const noexcept { return index >= 0 and index < m_rowsCount; }
   bool CheckColumnIndex(int index) const noexcept { return index >= 0 and index < m_columnsCount; }
 };
